숫자의 상태를 관리하는 Counter 컴포넌트를 추출한다
- App 컴포넌트에서 Counter 컴포넌트를 추출한다.
- Counter 컴포넌트는 숫자의 상태를 관리한다.
